requirements.txt
streamlit
langchain
faiss-cpu
pypdf
sentence-transformers
